worker/pyproject.toml
[project]
name = "screening-worker"
version = "0.1.0"
description = "Worker for recruitment screening system"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0.25",
    "asyncmy>=0.2.9",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "openai>=1.10.0",
    "pypdf>=3.17.4",
    "python-docx>=1.1.0",
    "aiofiles>=23.2.1",
    "numpy>=1.26.3",
    "python-magic>=0.4.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["worker"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
